# docs/openapi.yaml
openapi: 3.0.0
info:
  title: API Services
  version: v1.1
  description: This API provides all the necessary endpoints to give access to specific informations to Surirobot.
servers:
  - url: 'https://services.api.suricats-consulting.com/api'
    description: Main (production) server
  - url: 'https://services.api.surirobot.fr/api'
    description: Dev server
  - url: 'http://localhost:5000/api'
    description: Local server
paths:
  /weather:
    post:
      summary: Give weather for a given time and location
      operationId: weather
      tags:
        - Services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeatherDarkSkyRequest'
      responses:
        200:
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherDarkSkyResponse'
        400:
          description: Wrong parameter(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: API Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        503:
          description: Forecast API is not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    WeatherDarkSkyRequest:
      required:
        - latitude
        - longitude
        - time
        - language
      properties:
        latitude:
          oneOf:
            - type: string
            - type: number
              format: float
        longitude:
          oneOf:
            - type: string
            - type: number
              format: float
        time:
          oneOf:
            - type: string
            - type: number
              format: float
        language:
          type: string
          enum:
            - fr
            - en
    WeatherDarkSkyResponse:
      required:
        - apparentTemperature
        - cloudCover
        - humidity
        - precipIntensity
        - precipProbability
        - summary
        - temperature
      properties:
        currently:
          type: object
          properties:
            apparentTemperature:
              type: number
              format: float
            cloudCover:
              type: number
              format: float
            humidity:
              type: number
              format: float
            precipIntensity:
              type: number
              format: float
            precipProbability:
              type: number
              format: float
            summary:
              type: string
            temperature:
              type: number
              format: float
        daily:
          type: object
          properties:
            apparentTemperature:
              type: number
              format: float
            cloudCover:
              type: number
              format: float
            humidity:
              type: number
              format: float
            precipIntensity:
              type: number
              format: float
            precipProbability:
              type: number
              format: float
            summary:
              type: string
            temperature:
              type: number
              format: float
...
    Error:
      required:
        - code
        - msg
      properties:
        code:
          type: string
          enum:
            - missing_header
            - bad_header
            - api_error
            - external_api_error
            - invalid_credentials_error
            - operation_failed
            - missing_parameter
            - resource_not_found
            - bad_parameter
        msg:
          type: string
    ErrorResponse:
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'